Format component
Created all the necessary code, and CSS associated with the Format component. This was the final component needed to make the whole app functional. The user is prompted to add their name, card number, card expiration date in MM, YY format and the CVC numbers in the back of each credit or debit card. All of this data is functional in the means that as they type it will also show up the the card image and once the data submits successfully a thank you message is prompted. If the data is not valid and error will pop up.